a lot of things
renamed icons and made it so that it displays different weather icons based on the weather code.

added chart, that diesplays the temperature, humidity, and the windspeed.

Also small weathercards now display the temperature min and max and also shows the according weather icon.

Tomorrow add functionality to automatically renew the jwt token and to be able to add fields, sensors and sensorvalues